requirements.txt
discord.py
youtube_dl
youtube-search-python